requirements.txt
astroid==2.4.2
colorama==0.4.3
Django==2.2
django-excel-response==2.0.5
django-simple-history==2.11.0
et-xmlfile==1.0.1
isort==5.4.2
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==2.4.11
Pillow==7.2.0
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
